Adjusted UART and Current Calibration
Changed the UART ports in which data is being transmitted on and changed the handledef to meet the requirements of transmitting the data out the virtual com port. Adjusted the current calibration to meet the output that will be apparent when the current pump resistor is changed to 180 ohms opposed to 150.