Programming Routine Changed
Program rountine was changed as the previous method did not work correctly resulting in the fuse bytes being incorrectly programmed. The change does not include verification, but to correctly work should be implemented.
Given the programming does not require more than 64Kbytes to be written to the target board and the MAIT software is in sequential order there is no requirement to test the address location at this stage, if code was being written do different locations noted by the record type in the intel hex format this would be required to be changed. As such, when the programming reaches 256 bytes loaded before the page write, the page is written and the page count increased ready for the next page.

Tasks to complete for working programming:
- Memory verification, ensuring that programming was successful with the correct data programmed in the right place.
- Fuse bytes set correctly following programming
- Lock byte set and clock returned to the external CLK