Calibrating and Serialising Boards
Added the ability to calibrate, serialise, initialise and test latches on boards.
Improved the Latch test functionality as the adc can now be polled on any channel without all channels requiring to be read.
Serialise boards, use the serial number that was input through keypad or barcode scanner to individually serialise each board.
Initialise boards, the ability to set the boards to default configuration. Meaning once the board is successfully programmed it can be placed on the self as per standard.
Work required to make a serialisation and calibrate routine for each board. Only boards that can be calibrated right now is the single and dual as they operate on the first output port.
Improvements on basic flow of program were also implemented with the use of the memcpy functionality.