Improved Board Struct for better testing process
Changed the system that was currently implemented that used a variety of global variables to the use of a struct that is passed throughout the program to improve the functionality and makes more sense.
Improved the functionality of the design as more functions rely less upon each other. This also allowed for the imrpovement of the test.c file as the loading of the communication buffer can be slightly changed to a more readable routine for easier debugging at a later date
Future Imrpovements:
-Improved accuracy of the testing
-Introduction of state machine for the testing process
         # includes timeout functionality for if the program is held in a current state it retrys or exits
-Testing of the 4011/4012 boards
         # Requires level shifter on comms and programming to protect the test rig