Line and Page Handling
Functionality added to handle the reading in of the hex program lines, when the page buffer is full a page write is called followed by placing the remaining data into the page buffer. Functionality should be added to verify and write the last page of data when the end of file is received.