TestRig V1.11
Stable testing process, with updated state machine.
Stable release for 4271, 9371, 9352.
Testing for 4011, 4020 and 4220 required for full stability
management of states far more efficient compared to previous releases of firmware, speed improvements can be made throughout some states along with a couple of functions removing dependencies on while(1) loops to stop the level of blocking, especially in the programming section.
Improvements made to timers of the system, possibly updating the clk to 72/96Mhz to allow for faster programming when the buffer system is implemented.