Test Rig v1.13
Functionality improvements
Added new function to job straight to programming, so that if the system is in the idle state if the 6 button is pressed the testrig will program the target board imediately opposed to attempting to interogate the board before programming commenses.
Operation will continue as previous with the 1 press, however changes were made to the flow of the states with the system reinterogating following an initialisation to ensure that the network and module ID can be attained to communicate with the board.