Friday Commit
Slight changes to a variety of board tests, such as the single, dual and 4220, work needs to be complete to determine why the last test on the 4270 does not return valid results following the latch test. At this stage all responses are correct following the configuration of the final test.
Other changes were made to the code to account for higher input voltages with changes made to the resistor values on the input voltage checks.
Changes to a few variable names to make code more user friendly