TestRig V1.21
Stable SDI-12 and RS385 with random number generation to produce a variety of tests. True random number generation is used on the HAL layer to obtain these random values.
Few changes made to improve the stability of testing on the 935x and 401x boards.
More work required on UI and async testing following a failed latch test.