Friday Commit
Fixes made to a variety of timers, removing HAL_Delays whereever possible to improve the operation of the code. Improvements made to the scan loom capabilities with the ability to scan throughout the system, with limitations placed on the calibration latching and programming. As these routines in the statemachine are dependant on the MUX lines.
Overall stability improvements with a few flaws associated with the home screen shifting between states and the different modes that the testing should occur under