TestRig V1.22.2
Stability improved for loom checking and overall performance, issues with LCD and system getting stuck seemed to be removed as issues with the casting of LoomSel is fixed.
Quiting from programming and other areas through system improved. Fixes to when ScanLoom() are not run need to be completed to fix the quit functionality fully'.